Fixed bug: Escape not working when in Pause
