+++ Fightship Arena/Assets/Scriptables/Orchestration/Level 1/Spawns/W1-group-001.asset
@@ -17,3 +17,4 @@ MonoBehaviour:
   - {fileID: 11400000, guid: c28131626d0702b42983bde480a15649, type: 2}
   - {fileID: 11400000, guid: e69818da75ff851448828aeeb069a9e9, type: 2}
   - {fileID: 11400000, guid: 27f1a5b70c03b734e864538380b4a8dc, type: 2}
+  - {fileID: 11400000, guid: aa14055120f33ae438a4b2010c944bc9, type: 2}
