+++ Fightship Arena/Assets/Scriptables/Orchestration/Level 1/Spawns/W1-inf-004.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa14055120f33ae438a4b2010c944bc9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
