+++ Fightship Arena/Assets/Scripts/Managers/Menus/PauseMenuManager.cs
@@ -14,8 +14,6 @@ namespace FightShipArena.Assets.Scripts.Managers.Menus
 
         public IPauseMenuManager Core { get; protected set; }
 
-        private float _timeScale;
-
         void Awake()
         {
             Core = new PauseMenuManagerCore(this);
@@ -31,8 +29,6 @@ namespace FightShipArena.Assets.Scripts.Managers.Menus
         public void OnStart()
         {
             Core.OnStart();
-            _timeScale = Time.timeScale;
-            SetTimeScale();
         }
 
         public void OnAwake()
@@ -46,22 +42,11 @@ namespace FightShipArena.Assets.Scripts.Managers.Menus
         public void ResumeGame()
         {
             Core.ResumeGame();
-            ResetTimeScale();
         }
 
         public void QuitCurrentGame()
         {
             Core.QuitCurrentGame();
-            ResetTimeScale();
-        }
-
-        private void SetTimeScale()
-        {
-            Time.timeScale = 0;
-        }
-        private void ResetTimeScale()
-        {
-            Time.timeScale = _timeScale;
         }
     }
 }
